requirements.txt
Django==4.2.11
Pillow==10.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Markdown==3.4.3
drf-yasg==1.21.5
django-widget-tweaks==1.5.0
django-admin-material-dashboard==1.0.10